requirements.txt
# RAG System Dependencies
# Core RAG components
sentence-transformers==2.2.2
faiss-cpu==1.7.4
langchain==0.1.0
langchain-community==0.0.10

# Document processing
pypdf==3.17.4
python-docx==1.1.0
beautifulsoup4==4.12.2

# LLM providers (OpenAI as default, add others as needed)
openai==1.6.1

# Utilities
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1
python-dotenv==1.0.0

# Analysis and visualization
matplotlib==3.8.2
seaborn==0.13.0
tabulate==0.9.0

# MCP SDK (for MCP-enhanced version)
mcp==0.1.0

# Testing
pytest==7.4.3
